# backend-api/artifacts/openapi/admin.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Metrial BaseCode API",
        "version": "v1",
        "description": "Metrial BaseCode is a tenant-aware business platform API for authentication, payments, wallets, media, governance, territories, currencies, integrations, and webhooks."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "SanctumBearer": []
        }
    ],
    "paths": {
...
        "/governance/settings/{key}": {
            "get": {
                "operationId": "governance.settings.show",
                "description": "Requires a Laravel Sanctum bearer token. Requires the permission declared by the route middleware.",
                "summary": "Retrieve resource details",
                "tags": [
                    "Governance"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "success": true,
                                            "message": null,
                                            "data": {
                                                "status": "completed"
                                            },
                                            "meta": {
                                                "request_id": "018f68c5-80fd-7b63-b56a-3aecc84f8b62",
                                                "locale": "en",
                                                "direction": "ltr"
                                            }
                                        }
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "examples": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "examples": [
                                                null
                                            ]
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "request_id": {
                                                    "type": "string",
                                                    "examples": [
                                                        "018f68c5-80fd-7b63-b56a-3aecc84f8b62"
                                                    ]
                                                },
                                                "locale": {
                                                    "type": "string",
                                                    "examples": [
                                                        "en"
                                                    ]
                                                },
                                                "direction": {
                                                    "type": "string",
                                                    "examples": [
                                                        "ltr"
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "governance.settings.update",
                "description": "Requires a Laravel Sanctum bearer token. Requires the permission declared by the route middleware.",
                "summary": "Update this resource",
                "tags": [
                    "Governance"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "success": true,
                                            "message": null,
                                            "data": {
                                                "status": "completed"
                                            },
                                            "meta": {
                                                "request_id": "018f68c5-80fd-7b63-b56a-3aecc84f8b62",
                                                "locale": "en",
                                                "direction": "ltr"
                                            }
                                        }
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "examples": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "examples": [
                                                null
                                            ]
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "request_id": {
                                                    "type": "string",
                                                    "examples": [
                                                        "018f68c5-80fd-7b63-b56a-3aecc84f8b62"
                                                    ]
                                                },
                                                "locale": {
                                                    "type": "string",
                                                    "examples": [
                                                        "en"
                                                    ]
                                                },
                                                "direction": {
                                                    "type": "string",
                                                    "examples": [
                                                        "ltr"
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "governance.settings.destroy",
                "description": "Requires a Laravel Sanctum bearer token. Requires the permission declared by the route middleware.",
                "summary": "Delete or revoke this resource",
                "tags": [
                    "Governance"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/webhook-endpoints/{webhookEndpoint}": {
            "put": {
                "operationId": "webhook.endpoints.update",
                "description": "Requires a Laravel Sanctum bearer token. Requires the permission declared by the route middleware.",
                "summary": "Update this resource",
                "tags": [
                    "Webhooks"
                ],
                "parameters": [
                    {
                        "name": "webhookEndpoint",
                        "in": "path",
                        "required": true,
                        "description": "The webhook endpoint ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWebhookEndpointRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "success": true,
                                            "message": null,
                                            "data": {
                                                "status": "completed"
                                            },
                                            "meta": {
                                                "request_id": "018f68c5-80fd-7b63-b56a-3aecc84f8b62",
                                                "locale": "en",
                                                "direction": "ltr"
                                            }
                                        }
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "examples": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "examples": [
                                                null
                                            ]
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "request_id": {
                                                    "type": "string",
                                                    "examples": [
                                                        "018f68c5-80fd-7b63-b56a-3aecc84f8b62"
                                                    ]
                                                },
                                                "locale": {
                                                    "type": "string",
                                                    "examples": [
                                                        "en"
                                                    ]
                                                },
                                                "direction": {
                                                    "type": "string",
                                                    "examples": [
                                                        "ltr"
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "webhook.endpoints.destroy",
                "description": "Requires a Laravel Sanctum bearer token. Requires the permission declared by the route middleware.",
                "summary": "Delete or revoke this resource",
                "tags": [
                    "Webhooks"
                ],
                "parameters": [
                    {
                        "name": "webhookEndpoint",
                        "in": "path",
                        "required": true,
                        "description": "The webhook endpoint ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "SanctumBearer": {
                "type": "http",
                "description": "Laravel Sanctum bearer token.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        },
        "schemas": {
...
            "StoreWebhookEndpointRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "maxLength": 100
                        },
                        "minItems": 1
                    },
                    "active": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "required": [
                    "name",
                    "url",
                    "events"
                ],
                "title": "StoreWebhookEndpointRequest"
            },
            "UpdateSettingRequest": {
                "type": "object",
                "properties": {
                    "value": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "any JSON value",
                        "maxLength": 500
                    }
                },
                "required": [
                    "value"
                ],
                "title": "UpdateSettingRequest"
            }
        },
        "responses": {
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}